파일 정리
폴더별로 파일 정리
