+++ client/public/index.html
@@ -6,6 +6,8 @@
       src="https://static.nid.naver.com/js/naveridlogin_js_sdk_2.0.0.js"
       charset="utf-8"
     ></script>
+    <link rel="icon" href="%PUBLIC_URL%/logo.png" />
+    <title>Tory's Journey</title>
   </head>
   <body>
     <div id="root"></div>
