--- client/src/Oauth/KakaoOauth.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-const KakaoOauth: React.FC = () => {
-  return (
-    <div>
-      <a href="/auth/kakao">카카오 간편 로그인</a>
-    </div>
-  );
-};
-
-export default KakaoOauth;
-
-//	b23facc67215c6c8e0f6e042bd040985
