+++ client/src/component/DetailCarousel.tsx
@@ -1,7 +1,5 @@
 import "react-image-gallery/styles/css/image-gallery.css";
 import ImageGallery from "react-image-gallery";
-import dummyData from "../pages/TESTTESTTEST";
-import { useParams } from "react-router-dom";
 import { CarouselButton, ImgModal } from "../style/RestaurantInfo";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faTimes } from "@fortawesome/free-solid-svg-icons";
