+++ client/src/component/Navbar.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { Link } from "react-router-dom";
 import { Menu, MenuItem, MenuLink } from "../style/Navbar";
 
 const Navbar = () => {
