--- client/src/modal/PictureDetail.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const PictureDetail = () => {
-  return <div>상세 사진 모음</div>;
-};
-
-export default PictureDetail;
