--- client/src/modal/PictureUpload.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const PictureUpload = () => {
-  return <div>PictureUpload</div>;
-};
-
-export default PictureUpload;
