+++ client/src/modal/SignUp.tsx
@@ -1,6 +1,5 @@
 import React, { useState } from "react";
 import axios from "axios";
-import { useNavigate } from "react-router-dom";
 import {
   SignUpModal,
   FormModal,
