+++ client/src/modal/WritingManage/WritingChangeModalOnModal.tsx
@@ -5,39 +5,34 @@ import {
   ChangeModal,
   Content,
   ButtonDiv2,
-  Close,
-  Button,
-} from "../style/Modal";
+} from "../../style/Modal";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faTimes } from "@fortawesome/free-solid-svg-icons";
-
-type modalChangeProps = {
+type valueProps = {
   modalChange: () => void;
-  patchReview: () => void;
+  patchWriting: () => void;
 };
 
-const ReviewChangeModalOnModal = ({
+const WritingChangeModalOnModal = ({
   modalChange,
-  patchReview,
-}: modalChangeProps) => {
+  patchWriting,
+}: valueProps) => {
   return (
     <ModalBack>
       <ChangeModal>
         <Header>
           <div>Tory's-journey</div>
-         <Close>
-          <FontAwesomeIcon onClick={modalChange} icon={faTimes} /></Close>
-
+          <FontAwesomeIcon onClick={modalChange} icon={faTimes} />
         </Header>
 
         <Content>정말 수정하시겠습니까?</Content>
         <ButtonDiv2>
-          <Button onClick={patchReview}>확인</Button> 
-          <Button onClick={modalChange}>취소</Button>
+          <button onClick={patchWriting}>확인</button>
+          <button onClick={modalChange}>취소</button>
         </ButtonDiv2>
       </ChangeModal>
     </ModalBack>
   );
 };
 
-export default ReviewChangeModalOnModal;
+export default WritingChangeModalOnModal;
