+++ client/src/pages/Main.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { Link } from "react-router-dom";
 import store from "../redux/Store";
 
 import {
