+++ client/src/pages/MyPage/MyReview.tsx
@@ -1,9 +1,9 @@
 import axios from "axios";
 import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import MyPageMenu from "../component/MyPageMenu";
-import MyReviewList from "../component/MyReviewList";
-import { MyPageDiv, ReviewDiv, Title } from "../style/Mypage";
+import MyPageMenu from "../../component/MyPageMenu";
+import MyReviewList from "../../component/MyReviewList";
+import { MyPageDiv, ReviewDiv, Title } from "../../style/Mypage";
 
 const MyReview = () => {
   const navigate = useNavigate();
