--- client/src/pages/Writing.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Writing = () => {
-  return (
-    <div>
-      <div>Writing</div>;
-    </div>
-  );
-};
-
-export default Writing;
